Move a ProposalStatusIcon to Proposals directory
(We want to group related components together)
